fix compiling issue after restructuring and refine dependency
