+++ app/website/pom.xml
@@ -39,13 +39,12 @@
   <dependencies>
     <!-- war dependencies -->
     <dependency>
-      <groupId>sjph.life.model</groupId>
-      <artifactId>sjph-life-model</artifactId>
-      <scope>runtime</scope>
-    </dependency>
-    <dependency>
       <groupId>sjph.life.platform</groupId>
       <artifactId>sjph-life-platform</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>sjph.life.model</groupId>
+      <artifactId>sjph-life-model</artifactId>
       <scope>runtime</scope>
     </dependency>
     <dependency>
