+++ platform/src/main/java/sjph/life/platform/cache/impl/RedisCacheServiceImpl.java
@@ -3,6 +3,7 @@ package sjph.life.platform.cache.impl;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
@@ -18,6 +19,7 @@ public class RedisCacheServiceImpl implements CacheService {
     private static final Logger LOGGER = LogManager.getLogger(RedisCacheServiceImpl.class);
     // inject the actual template
     @Autowired(required = true)
+    @Qualifier("redisTemplate")
     private RedisTemplate<String, String>  template;
 
     // inject the template as ListOperations
